+++ packages/react-article-components/src/constants/position-z-index.js
@@ -1,0 +1,4 @@
+export default {
+  // table-of-contents
+  toc: 10,
+}