+++ packages/react-article-components/webpack.config.js
@@ -12,7 +12,7 @@ const config = {
   },
   devServer: {
     hot: true,
-    watchContentBase: true,
+    watchContentBase: false,
     host: '0.0.0.0',
     port: 8080,
   },